fix: resolve issue in PR #5
